blog-work/ndcg-score/requirements.txt
scikit-surprise==1.1.4
numpy==1.24.4
scikit-learn==1.6.1
pandas==1.4.1
pyspark==3.2.1
findspark==2.0.1
psutil==7.0.0